requirements.txt
langgraph
langchain
langchain_openai
openai
tabulate
python-dotenv